From https://www.jwz.org/xscreensaver/xscreensaver-6.03.tar.gz
-rw-rw-r-- 1 zblaxell zblaxell 25703584 Feb 28 11:57 xscreensaver-6.03.tar.gz
04b272b7b4eadb68e812c682148162bd59d3b9f8  xscreensaver-6.03.tar.gz

+++ driver/test-screens.c
@@ -83,7 +83,7 @@ test (int testnum, const char *screens, const char *desired)
         {
           sprintf (out, "%dx%d+%d+%d", m->width, m->height, m->x, m->y);
           if (m->screen)
-            sprintf (out + strlen(out), "@%d", (int) m->screen);
+            sprintf (out + strlen(out), "@%ld", (long) m->screen);
         }
       else
         strcpy (out, failstr (m->sanity));
